+++ gantry_cia402/src/driver/oms.rs
@@ -96,7 +96,7 @@ pub struct VelocitySetpoint {
 #[derive(Clone, Debug)]
 pub struct TorqueSetpoint {
     // TODO uom?
-    pub target_torque: i32,
+    pub target_torque: i16,
 }
 
 pub struct OmsHandler {
